(fix) calendar item removes from calendar when favorite event is deleted from the list
